requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiocache==0.12.3
    # via ezcord
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   ezcord
    #   py-cord
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.22.1
    # via ezcord
attrs==25.4.0
    # via aiohttp
audioop-lts==0.2.2
    # via ezcord
colorama==0.4.6
    # via ezcord
ezcord==0.7.4
    # via cookiebot (pyproject.toml)
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.11
    # via yarl
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
py-cord==2.7.1
    # via cookiebot (pyproject.toml)
python-dotenv==1.2.1
    # via ezcord
pyyaml==6.0.3
    # via ezcord
typing-extensions==4.15.0
    # via py-cord
yarl==1.22.0
    # via aiohttp
